requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0